Set isbn retrieve to process in parallel
With these changes, the retrieve of isbn in crawler gateway is now parallel, increasing
performance. The strategy pattern implementation is now a  little different but it was
needed to have multiple strategies at the same time. In the older
version, the IsbnContext could have only one strategy, so it wasn't possible to
process in parallel, because each book can have it's own strategy and
change the strategy could lead to not desired results. Now as
IsbnContext doesn't have a setter to strategy, the client of the context
set the strategy per call on getIsbn, so each call can have a different
strategy and a call to getIsbn doesn't change the strategy to other
calls