Refact server to ease integration tests
With this commit, now it's possible to start multiple servers, so in
test we can start two different servers and test some scenarios, e.g.
the test with empty list returned by the crawler was not possible
because the site will always return a list with books, now is possible
to start the server but execute this test. The test was added into this
commit