Add cache behavior to crawler
Now is possible to use a cache feature in crawler, so the results is
much faster than the results without cache. The user has two options,
cache in startup, that will make the application startup take more time,
but all requests will be faster or only cache, that will happen only
after the first call, so the first call will be slower, but the others
will be faster and the startup time will be slower.